% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_flother.R
\name{theme_flother}
\alias{theme_flother}
\title{A ggplot2 theme in the style used on https://flother.is/}
\usage{
theme_flother(font_family = "Valkyrie T4", font_size = 12)
}
\arguments{
\item{font_family}{Font family for all text, by default Matthew Butterick's
Valkyrie (see \url{https://practicaltypography.com/valkyrie.html}).}

\item{font_size}{Base size of font. The size of all text, lines, and margins
is relative to this.}
}
\value{
A complete ggplot2 theme.
}
\description{
A ggplot2 theme in the style used on https://flother.is/
}
\examples{
library(ggplot2)
ggplot(USArrests, aes(Murder, UrbanPop)) +
  geom_point() +
  theme_flother()
}
